refactor: Refactored data and used barrel exports
Refactored the following:
1. Separated the data used in the components in a separate file
2. Used barrel exports
3. Replaced strings used in reducers as constants
