+++ src/Components/CategoryCard/StyleCategoryCard.jsx
@@ -1,13 +1,14 @@
 import "./category-card.css";
-import { useProducts } from "../../context/ProductContext";
+import { useProducts } from "../../context";
 import { useNavigate } from "react-router-dom";
+import { FILTER_BY_BRANDS } from "../../helper/constants";
 
 const StyleCategoryCard = ({ source, information, caption }) => {
   const { productsDispatch } = useProducts();
   const navigate = useNavigate();
   const showFilteredCategory = (title) => {
     productsDispatch({
-      type: "FILTER_BY_BRANDS",
+      type: FILTER_BY_BRANDS,
       payload: title,
     });
     navigate("/product");
