+++ src/Components/Navbar/Navbar.jsx
@@ -22,13 +22,22 @@ const Navbar = () => {
     <header>
       <nav className="navigation">
         <div className="nav-brand-logo">
-          <NavLink to="/">
+          <NavLink
+            to="/"
+            onClick={() => productsDispatch({ type: "CLEAR_ALL" })}
+          >
             <img className="nav-img" src={logo} alt="website logo" />
           </NavLink>
         </div>
         <div className="nav-main-options">
           <ul className="nav-ul">
-            <li className="nav-li" onClick={() => filterGender("men")}>
+            <li
+              className="nav-li"
+              onClick={() => {
+                productsDispatch({ type: "CLEAR_ALL" });
+                filterGender("men");
+              }}
+            >
               <NavLink
                 className={(navData) => (navData.isActive ? "nav-active" : "")}
                 to="/product"
@@ -36,7 +45,13 @@ const Navbar = () => {
                 MEN
               </NavLink>
             </li>
-            <li className="nav-li" onClick={() => filterGender("women")}>
+            <li
+              className="nav-li"
+              onClick={() => {
+                productsDispatch({ type: "CLEAR_ALL" });
+                filterGender("women");
+              }}
+            >
               <NavLink
                 to="/product"
                 className={(navData) => (navData.isActive ? "nav-active" : "")}
@@ -44,7 +59,13 @@ const Navbar = () => {
                 WOMEN
               </NavLink>
             </li>
-            <li className="nav-li" onClick={() => filterGender("children")}>
+            <li
+              className="nav-li"
+              onClick={() => {
+                productsDispatch({ type: "CLEAR_ALL" });
+                filterGender("children");
+              }}
+            >
               <NavLink
                 to="/product"
                 className={(navData) => (navData.isActive ? "nav-active" : "")}
@@ -66,7 +87,10 @@ const Navbar = () => {
         </div>
         <div className="nav-sub-options">
           <ul className="nav-ul">
-            <li className="nav-li">
+            <li
+              className="nav-li"
+              onClick={() => productsDispatch({ type: "CLEAR_ALL" })}
+            >
               <NavLink
                 to="/profile"
                 className={(navData) => (navData.isActive ? "nav-active" : "")}
@@ -74,7 +98,10 @@ const Navbar = () => {
                 <i className="fas fa-user"></i>
               </NavLink>
             </li>
-            <li className="nav-li">
+            <li
+              className="nav-li"
+              onClick={() => productsDispatch({ type: "CLEAR_ALL" })}
+            >
               <NavLink
                 to="/wishlist"
                 className={(navData) => (navData.isActive ? "nav-active" : "")}
@@ -85,7 +112,10 @@ const Navbar = () => {
                 )}
               </NavLink>
             </li>
-            <li className="nav-li">
+            <li
+              className="nav-li"
+              onClick={() => productsDispatch({ type: "CLEAR_ALL" })}
+            >
               <NavLink
                 to="/cart"
                 className={(navData) => (navData.isActive ? "nav-active" : "")}
