+++ src/helper/constants.js
@@ -1,22 +1,40 @@
+// Authentication constants
 export const AUTHLOADING = "AUTHLOADING";
 export const LOGIN = "LOGIN";
 export const SIGNUP = "SIGNUP";
 export const LOGOUT = "LOGOUT";
 export const AUTHERROR = "AUTHERROR";
+
+// Cart Management constants
 export const GET_CART_ITEMS = "GET_CART_ITEMS";
 export const ADD_TO_CART = "ADD_TO_CART";
 export const DELETE_FROM_CART = "DELETE_FROM_CART";
 export const CHANGE_CART_QUANTITY = "CHANGE_CART_QUANTITY";
 export const CART_LOADING = "CART_LOADING";
 export const CART_ERROR = "CART_ERROR";
+
+// Wishlist Management contants
 export const GET_WISHLIST_ITEMS = "GET_WISHLIST_ITEMS";
 export const ADD_TO_WISHLIST = "ADD_TO_WISHLIST";
 export const DELETE_FROM_WISHLIST = "DELETE_FROM_WISHLIST";
 export const WISHLIST_LOADING = "WISHLIST_LOADING";
 export const WISHLIST_ERROR = "WISHLIST_ERROR";
+
+// Product constants
 export const GET_ALL_PRODUCTS = "GET_ALL_PRODUCTS";
 export const PRODUCT_LOADING = "PRODUCT_LOADING";
 export const PRODUCT_ERROR = "PRODUCT_ERROR";
+
+// Single product constants
 export const GET_SINGLE_PRODUCT = "GET_SINGLE_PRODUCT";
 export const GET_SINGLE_PRODUCT_LOADING = "GET_SINGLE_PRODUCT_LOADING";
 export const GET_SINGLE_PRODUCT_ERROR = "GET_SINGLE_PRODUCT_ERROR";
+
+// Filter constants
+export const FILTER_BY_PRICE = "FILTER_BY_PRICE";
+export const FILTER_BY_BRANDS = "FILTER_BY_BRANDS";
+export const RATING = "RATING";
+export const PRICE_HIGH_TO_LOW = "PRICE_HIGH_TO_LOW";
+export const PRICE_LOW_TO_HIGH = "PRICE_LOW_TO_HIGH";
+export const FILTER_BY_GENDER = "FILTER_BY_GENDER";
+export const CLEAR_ALL = "CLEAR_ALL";
