+++ src/reducer/filterReducer.js
@@ -1,15 +1,24 @@
 import { initialState } from "../helper";
-import { GET_ALL_PRODUCTS } from "../helper/constants";
+import {
+  GET_ALL_PRODUCTS,
+  FILTER_BY_PRICE,
+  FILTER_BY_BRANDS,
+  RATING,
+  PRICE_HIGH_TO_LOW,
+  PRICE_LOW_TO_HIGH,
+  FILTER_BY_GENDER,
+  CLEAR_ALL,
+} from "../helper/constants";
 
 const filterReducer = (state, action) => {
   switch (action.type) {
     case GET_ALL_PRODUCTS:
       return { ...state, productData: action.payload };
 
-    case "FILTER_BY_PRICE":
+    case FILTER_BY_PRICE:
       return { ...state, pricerange: action.payload };
 
-    case "FILTER_BY_BRANDS":
+    case FILTER_BY_BRANDS:
       return {
         ...state,
         listofbrands: state.listofbrands.includes(action.payload)
@@ -17,19 +26,19 @@ const filterReducer = (state, action) => {
           : [...state.listofbrands, action.payload],
       };
 
-    case "RATING":
+    case RATING:
       return { ...state, rating: action.payload };
 
-    case "PRICE_HIGH_TO_LOW":
+    case PRICE_HIGH_TO_LOW:
       return { ...state, sorting: action.payload };
 
-    case "PRICE_LOW_TO_HIGH":
+    case PRICE_LOW_TO_HIGH:
       return { ...state, sorting: action.payload };
 
-    case "FILTER_BY_GENDER":
+    case FILTER_BY_GENDER:
       return { ...state, gender: action.payload };
 
-    case "CLEAR_ALL":
+    case CLEAR_ALL:
       return { ...initialState };
 
     default:
