+++ src/reducer/index.js
@@ -1,1 +1,3 @@
 export { authReducer } from "./authReducer";
+export { cartWishlistReducer } from "./cartWishlistReducer";
+export { filterReducer } from "./filterReducer";
